requirements.txt
pandas
openai
python-decouple
scipy